fix: shifted error handling
