+++ libs/query/src/lib/aa/stats/stats.service.ts
@@ -4,6 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useProjectAction } from '../../projects';
 import { useStatsStore } from './stats.store';
 import { UUID } from 'crypto';
+import { useSwal } from 'libs/query/src/swal';
 
 export const usePhasesStats = (uuid: UUID) => {
   const q = useProjectAction();
@@ -145,20 +146,38 @@ export const useProjectDashboardReporting = (uuid: UUID) => {
 
 export const useTransportSessionStats = (uuid: UUID) => {
   const q = useProjectAction();
+  const alert = useSwal();
+  const toast = alert.mixin({
+    toast: true,
+    position: 'top-end',
+    showConfirmButton: false,
+    timer: 3000,
+  });
 
   const query = useQuery({
     queryKey: ['transportSessionStats', uuid],
     queryFn: async () => {
-      const mutate = await q.mutateAsync({
-        uuid,
-        data: {
-          action: 'ms.activities.communication.getTransportSessionStats',
-          payload: {
-            projectId: uuid,
+      try {
+        const mutate = await q.mutateAsync({
+          uuid,
+          data: {
+            action: 'ms.activities.communication.getTransportSessionStats',
+            payload: {
+              projectId: uuid,
+            },
           },
-        },
-      });
-      return mutate.data;
+        });
+        return mutate.data;
+      } catch (error: any) {
+        const errorMessage =
+          error?.response?.data?.message || 'Failed to fetch transport stats';
+        toast.fire({
+          title: 'Error loading transport stats',
+          text: errorMessage,
+          icon: 'error',
+        });
+        throw error;
+      }
     },
     staleTime: 60 * 60 * 1000,
   });
